requirements.txt
imutils
opencv-python
ultralytics
pygame
Pillow

customtkinter
